fix: corrige tipo do id de usuario retornado pelo banco de plataformas (bigint vira string)
O pool pg retorna colunas bigint/int8 como string em JS por padrao,
mesmo com a interface PlatformUserRow declarando id: number. Isso
fazia createAdminSessionToken assinar um token com userId string, e
verifyAdminSessionToken rejeitar esse mesmo token no proximo request
(typeof payload.userId !== 'number'), gerando loop de redirecionamento
para /admin/login logo apos um login valido.

Cast id::integer na query resolve na fonte, sem mexer na configuracao
global de type-parser do pg (que afetaria tambem o pool de lib/db.ts).

+++ lib/platforms-db.ts
@@ -35,7 +35,7 @@ export interface PlatformUserRow {
 
 export async function findPlatformUserByUsername(username: string): Promise<PlatformUserRow | null> {
   const result = await platformsPool.query<PlatformUserRow>(
-    `SELECT id, username, name, hash, platforms
+    `SELECT id::integer, username, name, hash, platforms
      FROM users
      WHERE LOWER(username) = LOWER($1)
      LIMIT 1`,
